just finished adding safe args to the project.
ended up using serializable instead of parcelable <- did this due to it is how i am used to working with passing classes / data with intents.

first navigation action moves a generic value in this case a string to the next fragment.
second navigation action moves a class to the next fragment. (implementing Serializable on class and selecting within safe arg nav graph creation)